==> 1.1 prep_sample_info (Python).ipynb <==
#### This notebook is to create the preparation information and sample information files for Qiita, as well as the mapping file for Qiime.

import numpy as np
import pandas as pd

## create preparation_info document

info = pd.read_csv('../data/Demultiplex_Sheet.txt', sep='\t')

info = info.sort_values(['#SampleID'])
info.head()

info = info.drop_duplicates(subset=['#SampleID', 'ReversePrimer'])

info.head()

info.shape[0]

n = info.shape[0]
data = {'#SampleID': info.loc[:,'#SampleID'].str[7:13].values, # .values takes only value not index
       'Barcode': info.loc[:,'BarcodeSequence'].values,
       'LinkerPrimerSequence': info.loc[:, 'LinkerPrimerSequence'].values,
       'Description': info.loc[:,'Description'].values,
       'Experiment_Design_Description': np.repeat(
           '16S stool samples sequenced for MrOS Vitamin D study', n),
       'Library_Construction_Protocol': np.repeat('16S rRNA v4', n),
       'Linker': np.repeat('GT', n),
       'Platform': np.repeat('Illumina', n),
       'Center_Name': info.loc[:,'#SampleID'].str[14:16].values,
       'Center_Project': np.repeat('MrOS', n),
       'Instrument_Model': np.repeat('Illumina MiSeq', n)}
prep_info = pd.DataFrame(data, columns=['#SampleID', 'Barcode', 'LinkerPrimerSequence', 
                                        'Description', 
                                        'Experiment_Design_Description',
                                        'Library_Construction_Protocol', 'Linker', 'Platform',
                                        'Center_Name', 'Center_Project', 'Instrument_Model'])

prep_info.shape

prep_info.head(10)

# fix mismatch sample name 'BI0778' should be 'BIO778' (Nora's email on08/07/2017)
prep_info = prep_info.replace(to_replace='BIO778', value='BI0778')

## Create sample_info document

# note AMPHIT_15SD was missing in Nov6th metadata, thus copied from file Variable_sleepV4updated.csv

samples = pd.read_csv('../data/SleepMetadataV4_Nov6th2017.csv', sep=',')

samples.shape

# error in categories of SLEEPHRS (additional space in second '2')
samples.SLEEPHRS.value_counts()

# correct
samples.loc[samples.SLEEPHRS == '2: 7 HRS ', 'SLEEPHRS'] = '2: 7 HRS'

# check
samples.SLEEPHRS.value_counts()

samples.TUDRAMT.value_counts() # alcohol drinking variable is wrong

# replace alcohol drinking variable
tmp = pd.read_csv('../data/mapping_vitd_v4.txt', sep='\t')
tmp = tmp[['ID', 'TUDRAMT']]
tmp.rename(columns={'TUDRAMT':'TUDRAMT_REVISED'}, inplace=True)

tmp.shape

tmp.TUDRAMT_REVISED.value_counts()

tmp.head(3)

samples.head(3)

samples = pd.merge(samples, tmp, left_on='ID', right_on='ID')

samples.columns

samples.shape

len(pd.unique(samples.ID))

# reference: http://www.latlong.net/
Latitude = []
Longitude = []
sites = samples.loc[:, 'SITE']
for i in range(samples.shape[0]):
    if sites[i] == 'Birmingham':
        latitude = '33.520661'
        longitude = '-86.80249'
    elif sites[i] == 'San Diego':
        latitude = '32.715738'
        longitude = '-117.1611'
    elif sites[i] == 'Pittsburgh':
        latitude = '40.440625'
        longitude = '-79.99589'
    elif sites[i] == 'Palo Alto':
        latitude = '37.441883'
        longitude = '-122.143'
    elif sites[i] == 'Portland':
        latitude = '45.523062'
        longitude = '-122.6765'    
    elif sites[i] == 'Minneapolis':
        latitude = '44.977753'
        longitude = '-93.26501'  
    Latitude.append(latitude)
    Longitude.append(longitude)

# simple check
print(samples['SITE'][[1,90,200,300,400, 500]])
print(np.array(Latitude)[[1,90,200,300,400, 500]])
print(np.array(Longitude)[[1,90,200,300,400, 500]])

m = samples.shape[0]
required = {'#SampleID': samples.loc[:,'ID'].values,
        'Title': np.repeat('MrOS_VitaminD', m),
        'Anonymized_Name': samples.loc[:,'ID'].values,
        'Scientific_Name': np.repeat('human gut metagenome', m),
        'Taxon_ID': np.repeat('Not applicable', m),
        #'Description': np.repeat('Not applicable', m),
        'Sample_Type': np.repeat('stool', m),
        'Geo_Loc_Name': samples.loc[:, 'SITE'].values,
        'Elevation': np.repeat('Not applicable', m),
        'Env_Biome': np.repeat('urban biome', m),
        'Env_Feature': np.repeat('human-associated habitat', m),
        'Env_Material': np.repeat('feces', m),
        'Env_Package': np.repeat('human-gut', m),
        'Latitude': Latitude,
        'Longitude': Longitude,
        'Collection_Timestamp': np.repeat('Not applicable', m),
        'DNA_Extracted': np.repeat('Not applicable', m),
        'Physical_Specimen_Location': np.repeat('Not applicable', m),
        'Physical_Specimen_Remaining': np.repeat('Not applicable', m),
        'Age': samples.loc[:,'V4AGE1'].values,
        'Age_Units': np.repeat('years', m),
        'Host_Subject_ID': np.repeat('Not applicable', m),
        'Host_Taxid': np.repeat('Not applicable', m),
        'Host_Scientific_Name': np.repeat('Homo sapiens', m),
        'Host_Common_Name': np.repeat('human', m),
        'Life_Stage': np.repeat('adult', m),
        'Sex': np.repeat('male', m),
        'Height': samples.loc[:, 'HWHGT'].values,
        'Height_Units': np.repeat('cm', m),
        'Weight': samples.loc[:, 'HWWGT'].values,
        'Weight_Units': np.repeat('kg', m),  
        'BMI': samples.loc[:, 'HWBMI'].values,
        'Body_Habitat': np.repeat('UBERON:feces', m),
        'Body_Site': np.repeat('UBERON:feces', m),
        'Body_Product': np.repeat('UBERON:feces', m)}

len(required)

sample_info = pd.concat([pd.DataFrame(required), samples], axis=1)
sample_info = pd.DataFrame(sample_info, columns = ['#SampleID', 'Title', 'Anonymized_Name',
          'Scientific_Name', 'Taxon_ID', 
          #'Description', 
          'Sample_Type', 'Geo_Loc_Name', 'Elevation',
          'Env_Biome', 'Env_Feature', 'Env_Material', 'Env_Package', 'Latitude', 'Longitude',
          'Collection_Timestamp', 'DNA_Extracted', 'Physical_Specimen_Location', 'Physical_Specimen_Remaining',
          'Age', 'Age_Units', 'Host_Subject_ID', 'Host_Taxid', 'Host_Scientific_Name', 'Host_Common_Name',
          'Life_Stage', 'Sex', 'Height', 'Height_Units', 'Weight', 'Weight_Units', 'BMI', 'Body_Habitat',
          'Body_Site', 'Body_Product', 'GIERACE', 'SITE', 'AMAMPT', 'AMPHIT', 'AMFVT', 'MIDATA', 'V4DATE',
          'GIMSTAT', 'MHDIAB', 'MHRHEU1', 'MHOA', 'MHCHF', 'MHMI', 'MHDEPR', 'MHDEPRT', 'TUDRAMT_REVISED', 'PQPSLMED', 
          'PQPSQUAL', 'PASCORE', 'QLCOMP', 'SLSLPHRS', 'SLEEPHRS', 'PQPEFFCY', 'PQPEFFIC', 'PQPSQI', 'PQBADSLP', 'TURSMOKE', 
          'M1ADEPR', 'M1BENZO', 'AMAMPT_C1', 'AMFVT_C1', 'AMPHIT_15SD'])

sample_info.shape

sample_info.head(10)

## Match sample_name in two documents

id_prep=list(prep_info.loc[:,'#SampleID'].values)
id_sample=list(sample_info.loc[:,'#SampleID'].values)

# samples with metadata yet no sequencing data
for i in id_sample:
    if i not in id_prep:
        print(i)

# samples with sequencing data yet no metadata
no_sample_info = []
for j in id_prep:
    if j not in id_sample:
        no_sample_info.append(j)

print(len(no_sample_info))
print(no_sample_info)

## Drop un-needed samples

# exclude PO7100, as it has no microbiome data (see Lily Liu's email on 08/07/2017)
prep_info = prep_info.loc[~prep_info['#SampleID'].isin(['PO7100'])]
sample_info = sample_info.loc[~sample_info['#SampleID'].isin(['PO7100'])]

print(prep_info.shape)
print(sample_info.shape)

## create Qiita prep and Qiime mapping files

prep_info.head(5)

mapping = pd.merge(prep_info, sample_info, on='#SampleID')

print(prep_info.shape)
print(sample_info.shape)
print(mapping.shape)

#mapping.rename(columns={'Sample_Name': '#SampleID'}, inplace=True)
mapping.rename(columns={'Barcode': 'BarcodeSequence'}, inplace=True)

mapping.shape

mapping.head(5)

# move 'Description' to the end
cols = list(mapping)
cols.insert(mapping.shape[1], cols.pop(cols.index('Description')))
mapping = mapping.loc[:, cols]

mapping.head(5)

mapping.columns

# convert to proper variable types (categorical or continuous)
vars_cat = np.array(['BarcodeSequence', 'LinkerPrimerSequence', 'Experiment_Design_Description',
             'Library_Construction_Protocol', 'Linker', 'Platform', 'Center_Name', 'Center_Project', 'Instrument_Model',
             'Title', 'Anonymized_Name', 'Scientific_Name', 'Taxon_ID', 'Sample_Type', 'Geo_Loc_Name', 'Elevation', 'Env_Biome',
             'Env_Feature', 'Env_Material', 'Env_Package', 'Collection_Timestamp', 'DNA_Extracted', 'Physical_Specimen_Location',
             'Physical_Specimen_Remaining', 'Age_Units', 'Host_Subject_ID', 'Host_Taxid','Host_Scientific_Name', 'Host_Common_Name',
             'Life_Stage', 'Sex', 'Height_Units', 'Weight_Units', 'Body_Habitat', 'Body_Site', 'Body_Product', 'GIERACE', 'SITE',
             'SLEEPHRS', 'MIDATA', 'AMPHIT_15SD', 'MHDIAB', 'MHRHEU1', 'MHOA', 'MHCHF', 'MHMI', 'MHDEPR', 'GIMSTAT',
             'MHDEPRT', 'TUDRAMT_REVISED', 'PQPSQUAL', 'QLCOMP', 'PQBADSLP', 'TURSMOKE', 'M1ADEPR', 'M1BENZO', 'Description',
             'AMAMPT_C1', 'AMFVT_C1', 'V4DATE', 'PQPSLMED', 'PQPEFFIC'])
vars_cts = np.array(['Latitude', 'Longitude', 'Age', 'Height', 'Weight', 'BMI', 'PASCORE', 'AMAMPT', 
             'AMPHIT', 'AMFVT',  'PQPSQI', 'PQPEFFCY', 'SLSLPHRS'])

df = mapping.copy()
df[vars_cts] = df[vars_cts].apply(pd.to_numeric, errors='coerce')
df[vars_cat] = df[vars_cat].apply(lambda x: x.astype('category'))

prep_info.to_csv('../data/prep_sleep.txt', sep= '\t', index=False)
df.to_csv('../data/mapping_sleep.txt', sep= '\t', na_rep='Missing:not collected', index=False)
