==> 1.2 SEPP Tree, Alpha, Beta, Rarefy (Qiime2).ipynb <==
#### This notebook is to use Qiime2-2018.2 environment to construct phylogenetic tree, assign taxonomy, rarefaction, calculate alpha and beta diversities and generate PCoA plots

# compute alpha and beta diversity (include emperor plot for beta diversity)
Qm qiime diversity core-metrics-phylogenetic \
  --i-phylogeny insertion-tree.qza \
  --i-table mros_deblur_otus_unrare.qza \
  --p-sampling-depth 5807 \
  --m-metadata-file mapping_sleep_alpha.txt \
  --output-dir sleep-core-metrics-results
    
# export    
Qm qiime tools export \
  sleep-core-metrics-results/faith_pd_vector.qza \
  --output-dir sleep-alpha_PD

Qm qiime tools export \
  sleep-core-metrics-results/shannon_vector.qza \
  --output-dir sleep-alpha_Shannon

Qm qiime tools export \
  sleep-core-metrics-results/observed_otus_vector.qza \
  --output-dir sleep-alpha_ObsOTU

Qm qiime tools export \
  sleep-core-metrics-results/evenness_vector.qza \
  --output-dir sleep-alpha_Evenness

# taxonomy
Qm qiime taxa barplot \
--i-table mros_deblur_otus_unrare.qza \
--i-taxonomy taxonomy.qza \
--m-metadata-file mapping_sleep_alpha.txt \
--o-visualization taxonomy_sleep_barplot.qzv 

# ancom (no significant features detected)
Qm qiime taxa collapse \
  --i-table mros_deblur_otus_unrare.qza\
  --i-taxonomy taxonomy.qza \
  --p-level 6 \
  --o-collapsed-table sleep-table-l6.qza

Qm qiime composition add-pseudocount \
  --i-table sleep-table-l6.qza \
  --o-composition-table comp-sleep-table-l6.qza

Qm qiime composition ancom \
  --i-table comp-sleep-table-l6.qza \
  --m-metadata-file mapping_sleep_alpha.txt \
  --m-metadata-column PQBADSLP \
  --o-visualization l6-ancom-sleep-PQBADSLP.qzv
