==> 2.1 Alpha Diversity Analysis -- Add to mapping file (Python).ipynb <==
#### This notebook is to add alpha diversities to existing mapping file

import pandas as pd
import numpy as np

mf = pd.read_csv('../data/mapping_sleep.txt', sep='\t', dtype=str, index_col='#SampleID')

mf.head()

mf.shape

## add rarefied PD alpha diversity

alpha_PD = pd.read_csv('../Qiime_updated/alpha_PD/alpha-diversity.tsv', sep='\t', index_col=0)

alpha_PD.rename(columns={'faith_pd': 'alpha_pd'}, inplace=True)

alpha_PD.alpha_pd.describe()

df = pd.merge(mf, alpha_PD, left_index=True, right_index=True)
df.shape

## add Observed OTUs

alpha_OTU = pd.read_csv('../Qiime_updated/sleep-alpha_ObsOTU/alpha-diversity.tsv', sep='\t', index_col=0)

alpha_OTU.describe()

df = pd.merge(df, alpha_OTU, left_index=True, right_index=True)
df.shape

## add Shannon diversity

alpha_shannon = pd.read_csv('../Qiime_updated/alpha_Shannon/alpha-diversity.tsv', sep='\t', index_col=0)

df = pd.merge(df, alpha_shannon, left_index=True, right_index=True)
df.shape

## Pielou (Evenness)

alpha_evenness = pd.read_csv('../Qiime_updated/alpha_Evenness/alpha-diversity.tsv', sep='\t', index_col=0)

df = pd.merge(df, alpha_evenness, left_index=True, right_index=True)
df.shape

## export

### output mapping file with PD alpha diveristy

df.to_csv('../data/mapping_sleep_alpha.txt', sep='\t', index_label='#SampleID')
