==> 3.1 Beta Diversity (UW) -- RDA (R).ipynb <==
#### This notebook is to perform Non-Redundant Analysis on beta diversities (unweighted UniFrac)
#### reference: Falony G, Joossens M, Vieira-Silva S, et al. Population-level analysis of gut microbiome variation. Science 2016;352:560-4

library(vegan)

mf = read.csv('../data/mapping_sleep_alpha.txt', stringsAsFactors=FALSE, sep='\t')
colnames(mf)[1] = 'SampleName'
dim(mf)

pc = read.csv('../Qiime_updated/core-metrics-results/unweighted_unifrac_pcoa_results/ordination.txt', 
              header=FALSE, sep='\t', skip=9)

head(pc)

pc_10 = pc[, 1:11]
colnames(pc_10) = c('SampleName', 'PC1', 'PC2', 'PC3', 'PC4', 'PC5', 'PC6', 'PC7', 'PC8', 'PC9', 'PC10')
dim(pc_10)
head(pc_10)

head(mf)

dat = merge(mf, pc_10, by='SampleName')
dim(dat)

head(dat)

# convert categorial to factors
vars_cat = c('GIERACE', 'SITE', 'GIMSTAT', 'MHDIAB', 'MHRHEU1', 'MHOA', 'MHCHF', 'MHMI', 'MHDEPR', 
             'TUDRAMT_REVISED', 'PQPSLMED', 'PQPSQUAL', 'QLCOMP', 'SLEEPHRS', 'PQBADSLP', 
             'TURSMOKE', 'M1ADEPR', 'M1BENZO', 'PQPEFFIC', 'AMAMPT_C1', 'AMFVT_C1', 'AMPHIT_15SD')
dat[vars_cat] = lapply(dat[vars_cat], factor)

# # convert continuous to numeric
vars_cts = c('Age', 'BMI', 'PASCORE')
vars_pcs = c('PC1', 'PC2', 'PC3', 'PC4', 'PC5', 'PC6', 'PC7', 'PC8', 'PC9', 'PC10')

length(vars_cat)
length(vars_cts)
length(vars_pcs)

dat = dat[c(vars_cat, vars_cts, vars_pcs)]
dat = dat[complete.cases(dat), ]
X = dat[c(vars_cat, vars_cts)]
Y = dat[, vars_pcs]

dim(X)
dim(Y)

summary(X)

summary(Y)

mod0 <- rda(Y ~ 1., X)  # Model with intercept only
mod1 <- rda(Y ~ ., X)  # Model with all explanatory variables

step.res <- ordiR2step(mod0, mod1, perm.max = 1000)

table = step.res$anova
table

table$ES.RDA = c(table$R2.adj[1], table$R2.adj[2]-table$R2.adj[1], 
                 table$R2.adj[3]-table$R2.adj[2], table$R2.adj[4]-table$R2.adj[3],
                 table$R2.adj[5]-table$R2.adj[4], table$R2.adj[6]-table$R2.adj[5],
                 table$R2.adj[7]-table$R2.adj[6], table$R2.adj[8]-table$R2.adj[7])
table = table[-8, ]

step.res$call

table

library(ggplot2)

rownames(table) = c('Race', 'Diabetes', 'Antidepressent Use', 'Site', 'BMI',
                   'Poor Sleep', 'Smoking')

library(ggplot2)
pdf('../figures/RDA_beta_uw.pdf')
covariates = rownames(table)
ggplot(table, aes(x=reorder(covariates, ES.RDA), y=ES.RDA, fill=covariates)) +
  labs(x = 'Non-redundant Covariants', y = 'Effect Size (Unweighted UniFrac)') +
  geom_bar(stat='identity') +
 theme(axis.text=element_text(size=10), 
       axis.title=element_text(size=14,face="bold"), 
       legend.position="none") + 
  coord_flip()
dev.off()
