==> 3.3 Beta Diversity(Aitchison) -- Continuous (R).ipynb <==
#### This notebook is to perform Adonis test on continuous variables for beta diversities (Aitchison)

library(vegan)

dat <- read.delim("mapping_sleep_alpha.txt")
head(dat)

distance <- read.table(file = 'distance-matrix/distance-matrix.tsv', sep = '\t', header = TRUE, row.names = 1)
distance <- distance[ , order(colnames(distance))]
distance <- distance[ order(rownames(distance)), ]
distance[] <- lapply(distance, as.numeric)
head(distance)

#select test variables
varname <- c("X.SampleID", "Age", "AMAMPT", "AMPHIT", "AMFVT", "PQPSQI", "BMI", "Height", "Weight", "PASCORE", "SLSLPHRS", "PQPEFFCY")
meta <- dat[c(varname)]
head(meta)

lapply(meta, class)

set.seed(123)
meta$Age<- as.numeric(as.character(meta$Age))
age <- adonis(distance~Age, data = meta)
age

set.seed(123)
BMI<- adonis(distance~BMI, data = meta)
BMI

set.seed(123)
PASCORE<- adonis(distance~PASCORE, data = meta)
PASCORE

set.seed(123)
meta$AMAMPT <- as.numeric(as.character(meta$AMAMPT))
ind <- which(is.na(meta$AMAMPT))
dist <- distance[-ind, ]
dist <- dist[, -ind]
meta.omit <- meta[!is.na(meta$AMAMPT), ]
AMAMPT<- adonis(dist~AMAMPT, data = meta.omit)
AMAMPT

set.seed(123)
meta$SLSLPHRS <- as.numeric(as.character(meta$SLSLPHRS))
ind <- which(is.na(meta$SLSLPHRS))
dist <- distance[-ind, ]
dist <- dist[, -ind]
meta.omit <- meta[!is.na(meta$SLSLPHRS), ]
SLSLPHRS  <- adonis(dist~SLSLPHRS, data = meta.omit)
SLSLPHRS

set.seed(123)
meta$AMPHIT <- as.numeric(as.character(meta$AMPHIT))
ind <- which(is.na(meta$AMPHIT))
dist <- distance[-ind, ]
dist <- dist[, -ind]
meta.omit <- meta[!is.na(meta$AMPHIT), ]
acrophase  <- adonis(dist~AMPHIT, data = meta.omit)
acrophase

set.seed(123)
meta$AMFVT <- as.numeric(as.character(meta$AMFVT))
ind <- which(is.na(meta$AMFVT))
dist <- distance[-ind, ]
dist <- dist[, -ind]
meta.omit <- meta[!is.na(meta$AMFVT), ]
AMFVT  <- adonis(dist~AMFVT, data = meta.omit)
AMFVT

set.seed(123)
PQPSQI<- adonis(distance~PQPSQI, data = meta)
PQPSQI

set.seed(123)
meta$PQPEFFCY <- as.numeric(as.character(meta$PQPEFFCY))
ind <- which(is.na(meta$PQPEFFCY))
dist <- distance[-ind, ]
dist <- dist[, -ind]
meta.omit <- meta[!is.na(meta$PQPEFFCY), ]
PQPEFFCY  <- adonis(dist~PQPEFFCY, data = meta.omit)
PQPEFFCY

#### adjust p value by BH method

name <- c(  "Age", "BMI", "PASCORE", "AMAMPT", "SLSLPHRS", "AMPHIT", "AMFVT", "PQPSQI", "PQPEFFCY",
          'GIERACE', 'SITE', 'GIMSTAT', 'MHDIAB', 'MHRHEU1', 'MHOA', 'MHCHF', 'MHMI', 'MHDEPR', 
          'TUDRAMT_REVISED', 'PQPSLMED', 'PQPSQUAL', 'QLCOMP', 'SLEEPHRS', 'PQBADSLP', 'TURSMOKE', 
          'M1ADEPR', 'M1BENZO', 'PQPEFFIC', 'AMAMPT_C1', 'AMFVT_C1', 'AMPHIT_15SD')

pval <- c(0.09,0.035,0.001,0.482,0.411,0.628,0.001,0.01,0.018,
          0.001,0.006,0.1,0.001,0.562,0.038,0.829,0.886,0.031, 0.16,0.587,0.104,0.467,
          0.251,0.012,0.312,0.002,0.648,0.116,0.335, 0.004,0.137)

d <- cbind(name, pval)
d <-data.frame(name, pval, stringsAsFactors = F)
d <- d[-which(d$name %in% c('AMAMPT_C1', 'AMFVT_C1', 'AMPHIT_15SD','MHOA', 'MHCHF', 'MHMI','PQPSQUAL', 'QLCOMP')),]

d <- cbind(d, round(p.adjust(d$pval, "BH"), 4))
colnames(d)[3] <- "qval"
d
