==> 3.4 Beta Diversity(Aitchison) -- Control factors (R).ipynb <==
####  This notebook is to test for association between microbiome composition and sleep measurements controlling confounding factors (MiRKAT)

library(MiRKAT)

dat <- read.delim("mapping_sleep_alpha.txt")
head(dat)

distance <- read.table(file = 'distance-matrix.tsv', sep = '\t', header = TRUE, row.names = 1)
distance <- distance[ , order(colnames(distance))]
distance <- distance[ order(rownames(distance)), ]
distance[] <- lapply(distance, as.numeric)
head(distance)

vars_test = c('AMAMPT','AMPHIT','AMFVT','PQPEFFCY', 'PQPSQI','AMAMPT_C1' , 'AMFVT_C1','PQBADSLP')
vars_imp = c('GIERACE', 'SITE','MHDIAB', 'M1ADEPR', 'PASCORE',"BMI")
dat <- dat[c(vars_test,vars_imp)]
head(dat)

dat$AMFVT <- as.numeric(as.character(dat$AMFVT))
dat$PQPEFFCY <- as.numeric(as.character(dat$PQPEFFCY))

ind <- which(is.na(dat$AMFVT))
ind2 <- which(is.na(dat$PQPEFFCY))
dist <- distance[-c(ind,ind2), ]
dist <- dist[, -c(ind,ind2)]
dat <- na.omit(dat)
head(dat)

dim(dat)
head(dist)

test <- dat[c(vars_test)]
head(test)

imp <- dat[c(vars_imp)]
head(imp)

lapply(test, class)
lapply(imp, class)

imp[] <- lapply(imp, as.numeric)
test[, 1:5] <- lapply(test[, 1:5], as.character)
test[, 1:5] <- lapply(test[, 1:5], as.numeric)
test[, 6:8] <- lapply(test[, 6:8], as.factor)
test[, 6:8] <- lapply(test[, 6:8], as.numeric)
test[, 6:8] <- sapply(c(6:8), function(i){
    a <- as.numeric(test[, i])
    ifelse(a==2, 1, 0)})
head(test)

head(dist)

ker <- D2K(as.matrix(dist))

set.seed(10)
MiRKAT(test$AMFVT, imp, ker, out_type = "C", method = "permutation")
MiRKAT(test$AMFVT_C1, imp, ker, out_type = "D", method = "permutation")

set.seed(10)
MiRKAT(test$AMAMPT, imp, ker, out_type = "C", method = "permutation")
MiRKAT(test$AMAMPT_C1, imp, ker, out_type = "D", method = "permutation")

set.seed(10)
MiRKAT(test$AMPHIT, imp, ker, out_type = "C", method = "permutation")

set.seed(10)
MiRKAT(test$PQPEFFCY, imp, ker, out_type = "C", method = "permutation")
MiRKAT(test$PQBADSLP, imp, ker, out_type = "D", method = "permutation")
MiRKAT(test$PQPSQI, imp, ker, out_type = "C", method = "permutation")
