==> 4.1 Differential Analysis -- Missing Value (R).ipynb <==
#### This notebook is to handle missing values from data collection

meta <- read.delim("mapping_sleep_alpha.txt")
meta$AMPHIT <- as.numeric(as.character(meta$AMPHIT))
meta$AMFVT <- as.numeric(as.character(meta$AMFVT))
meta$SLSLPHRS<- as.numeric(as.character(meta$SLSLPHRS))

ind <- which(is.na(meta$AMPHIT))
ind2 <- which(is.na(meta$SLSLPHRS))
meta <- na.omit(meta)
colnames(meta)[1] <- "#SampleID"
head(meta)

write.table(meta, "mapping_sleep_alpha_omit.txt", sep = "\t", row.names = F)

meta.omit <- read.delim("mapping_sleep_alpha_omit.txt")
meta.omit<-meta.omit[!(meta.omit$SLSLPHRS==7),]
meta.omit$SLSLPHRS <- as.numeric(meta.omit$SLSLPHRS)
colnames(meta.omit)[1] <- "#SampleID"
head(meta.omit)

write.table(meta, "mapping_sleep_alpha_omit_sleep.txt", sep = "\t", row.names = F)
