==> 4.2 Differential Analysis -- Songbird (Qiime2).ipynb <==
### This notebook is to perform analysis on log-fold change of features with respect to measurements in metadata

!qiime dev refresh-cache

!qiime songbird multinomial \
    --i-table mros_deblur_otus_unrare.qza \
    --m-metadata-file mapping_sleep_alpha_omit.txt \
    --p-formula "AMAMPT+BMI+GIERACE+SITE+MHDIAB+M1ADEPR+GIMSTAT" \
    --p-epochs 1000 \
    --p-differential-prior 0.5 \
    --p-summary-interval 1 \
    --o-differentials Songbird/amplitute.qza \
    --o-regression-stats Songbird/regression-stats-amplitute.qza \
    --o-regression-biplot Songbird/regression-biplot-amplitute.qza
!qiime songbird summarize-single \
    --i-feature-table mros_deblur_otus_unrare.qza \
    --i-regression-stats Songbird/regression-stats-amplitute.qza \
    --o-visualization Songbird/regression-summary-amplitute.qzv
!qiime metadata tabulate \
    --m-input-file Songbird/amplitute.qza \
    --o-visualization Songbird/amplitute-viz.qzv
!qiime tools export \
  --input-path Songbird/amplitute-viz.qzv \
  --output-path Songbird/amplitute-table

!qiime songbird multinomial \
    --i-table mros_deblur_otus_unrare.qza \
    --m-metadata-file mapping_sleep_alpha_omit.txt \
    --p-formula "AMAMPT_C1+BMI+GIERACE+SITE+MHDIAB+M1ADEPR+GIMSTAT" \
    --p-epochs 500 \
    --p-differential-prior 0.5 \
    --p-summary-interval 1 \
    --o-differentials Songbird/amplitute-cat.qza \
    --o-regression-stats Songbird/regression-stats-amplitute-cat.qza \
    --o-regression-biplot Songbird/regression-biplot-amplitute-cat.qza
!qiime songbird summarize-single \
    --i-feature-table mros_deblur_otus_unrare.qza \
    --i-regression-stats Songbird/regression-stats-amplitute-cat.qza \
    --o-visualization Songbird/regression-summary-amplitute-cat.qzv
!qiime metadata tabulate \
    --m-input-file Songbird/amplitute-cat.qza \
    --o-visualization Songbird/amplitute-cat-viz.qzv
!qiime tools export \
  --input-path Songbird/amplitute-cat-viz.qzv \
  --output-path Songbird/amplitute-cat-table

!qiime songbird multinomial \
    --i-table mros_deblur_otus_unrare.qza \
    --m-metadata-file mapping_sleep_alpha_omit.txt \
    --p-formula "AMPHIT_15SD+BMI+GIERACE+SITE+MHDIAB+M1ADEPR+GIMSTAT" \
    --p-epochs 1000 \
    --p-differential-prior 0.5 \
    --p-summary-interval 1 \
    --o-differentials Songbird/AcrophaseSD.qza \
    --o-regression-stats Songbird/regression-stats-AcrophaseSD.qza \
    --o-regression-biplot Songbird/regression-biplot-AcrophaseSD.qza
!qiime songbird summarize-single \
    --i-feature-table mros_deblur_otus_unrare.qza \
    --i-regression-stats Songbird/regression-stats-AcrophaseSD.qza \
    --o-visualization Songbird/regression-summary-AcrophaseSD.qzv
!qiime metadata tabulate \
    --m-input-file Songbird/AcrophaseSD.qza \
    --o-visualization Songbird/AcrophaseSD-viz.qzv
!qiime tools export \
  --input-path Songbird/AcrophaseSD-viz.qzv \
  --output-path Songbird/AcrophaseSD-table

!qiime songbird multinomial \
    --i-table mros_deblur_otus_unrare.qza \
    --m-metadata-file mapping_sleep_alpha_omit.txt \
    --p-formula "AMPHIT+BMI+GIERACE+SITE+MHDIAB+M1ADEPR+GIMSTAT" \
    --p-epochs 500 \
    --p-differential-prior 0.5 \
    --p-summary-interval 1 \
    --o-differentials Songbird/Acrophase.qza \
    --o-regression-stats Songbird/regression-stats-Acrophase.qza \
    --o-regression-biplot Songbird/regression-biplot-Acrophase.qza
!qiime songbird summarize-single \
    --i-feature-table mros_deblur_otus_unrare.qza \
    --i-regression-stats Songbird/regression-stats-Acrophase.qza \
    --o-visualization Songbird/regression-summary-Acrophase.qzv
!qiime metadata tabulate \
    --m-input-file Songbird/Acrophase.qza \
    --o-visualization Songbird/Acrophase-viz.qzv
!qiime tools export \
  --input-path Songbird/Acrophase-viz.qzv \
  --output-path Songbird/Acrophase-table

!qiime songbird multinomial \
    --i-table mros_deblur_otus_unrare.qza \
    --m-metadata-file mapping_sleep_alpha_omit.txt \
    --p-formula "AMFVT+BMI+GIERACE+SITE+MHDIAB+M1ADEPR+GIMSTAT" \
    --p-epochs 1000 \
    --p-differential-prior 0.5 \
    --p-summary-interval 1 \
    --o-differentials Songbird/FStats.qza \
    --o-regression-stats Songbird/regression-stats-FStats.qza \
    --o-regression-biplot Songbird/regression-biplot-FStats.qza
!qiime songbird summarize-single \
    --i-feature-table mros_deblur_otus_unrare.qza \
    --i-regression-stats Songbird/regression-stats-FStats.qza \
    --o-visualization Songbird/regression-summary-FStats.qzv
!qiime metadata tabulate \
    --m-input-file Songbird/Fstats.qza \
    --o-visualization Songbird/Fstats-viz.qzv
!qiime tools export \
  --input-path Songbird/Fstats-viz.qzv \
  --output-path Songbird/Fstats-table

!qiime songbird multinomial \
    --i-table mros_deblur_otus_unrare.qza \
    --m-metadata-file mapping_sleep_alpha_omit.txt \
    --p-formula "AMFVT_C1+BMI+GIERACE+SITE+MHDIAB+M1ADEPR+GIMSTAT" \
    --p-epochs 1000 \
    --p-differential-prior 0.5 \
    --p-summary-interval 1 \
    --o-differentials Songbird/FStats-cat.qza \
    --o-regression-stats Songbird/regression-stats-FStats-cat.qza \
    --o-regression-biplot Songbird/regression-biplot-FStats-cat.qza
!qiime songbird summarize-single \
    --i-feature-table mros_deblur_otus_unrare.qza \
    --i-regression-stats Songbird/regression-stats-FStats-cat.qza \
    --o-visualization Songbird/regression-summary-FStats-cat.qzv
!qiime metadata tabulate \
    --m-input-file Songbird/Fstats-cat.qza \
    --o-visualization Songbird/Fstats-cat-viz.qzv
!qiime tools export \
  --input-path Songbird/Fstats-cat-viz.qzv \
  --output-path Songbird/Fstats-cat-table

!qiime songbird multinomial \
    --i-table mros_deblur_otus_unrare.qza \
    --m-metadata-file mapping_sleep_alpha_omit.txt \
    --p-formula "PQBADSLP+BMI+GIERACE+SITE+MHDIAB+M1ADEPR+GIMSTAT" \
    --p-epochs 1000 \
    --p-differential-prior 0.5 \
    --p-summary-interval 1 \
    --o-differentials Songbird/Poorslp.qza \
    --o-regression-stats Songbird/regression-stats-Poorslp.qza \
    --o-regression-biplot Songbird/regression-biplot-Poorslp.qza
!qiime songbird summarize-single \
    --i-feature-table mros_deblur_otus_unrare.qza \
    --i-regression-stats Songbird/regression-stats-Poorslp.qza \
    --o-visualization Songbird/regression-summary-Poorslp.qzv
!qiime metadata tabulate \
    --m-input-file Songbird/Poorslp.qza \
    --o-visualization Songbird/Poorslp-viz.qzv
!qiime tools export \
  --input-path Songbird/Poorslp-viz.qzv \
  --output-path Songbird/Poorslp-table

!qiime songbird multinomial \
    --i-table mros_deblur_otus_unrare.qza \
    --m-metadata-file mapping_sleep_alpha_omit.txt \
    --p-formula "SLEEPHRS+BMI+GIERACE+SITE+MHDIAB+M1ADEPR+GIMSTAT" \
    --p-epochs 1000 \
    --p-differential-prior 0.5 \
    --p-summary-interval 1 \
    --o-differentials Songbird/Slphours.qza \
    --o-regression-stats Songbird/regression-stats-Slphours.qza \
    --o-regression-biplot Songbird/regression-biplot-Slphours.qza
!qiime songbird summarize-single \
    --i-feature-table mros_deblur_otus_unrare.qza \
    --i-regression-stats Songbird/regression-stats-Slphours.qza \
    --o-visualization Songbird/regression-summary-Slphours.qzv
!qiime metadata tabulate \
    --m-input-file Songbird/Slphours.qza \
    --o-visualization Songbird/Slphours-viz.qzv
!qiime tools export \
  --input-path Songbird/Slphours-viz.qzv \
  --output-path Songbird/Slphours-table

!qiime songbird multinomial \
    --i-table mros_deblur_otus_unrare.qza \
    --m-metadata-file mapping_sleep_alpha_omit_sleep.txt \
    --p-formula "SLSLPHRS+BMI+GIERACE+SITE+MHDIAB+M1ADEPR+GIMSTAT" \
    --p-epochs 1000 \
    --p-differential-prior 0.5 \
    --p-summary-interval 1 \
    --o-differentials Songbird/Slphours-con.qza \
    --o-regression-stats Songbird/regression-stats-Slphours-con.qza \
    --o-regression-biplot Songbird/regression-biplot-Slphours-con.qza
!qiime songbird summarize-single \
    --i-feature-table mros_deblur_otus_unrare.qza \
    --i-regression-stats Songbird/regression-stats-Slphours-con.qza \
    --o-visualization Songbird/regression-summary-Slphours-con.qzv
!qiime metadata tabulate \
    --m-input-file Songbird/Slphours-con.qza \
    --o-visualization Songbird/Slphours-con-viz.qzv
!qiime tools export \
  --input-path Songbird/Slphours-con-viz.qzv \
  --output-path Songbird/Slphours-con-table

#Get plot of feature rankings
!qiime qurro loading-plot \
    --i-table mros_deblur_otus_unrare.qza \
    --i-ranks ordination.qza \
    --m-sample-metadata-file mapping_sleep_alpha_omit.txt \
    --m-feature-metadata-file taxonomy.qza \
    --o-visualization qurro-plot.qzv
