==> 4.3 Differentials Analysis-- Ranking (Python).ipynb <==
#### This notebook is to rank differentials for sleep measurements

import pandas as pd
import numpy as np

tax = pd.read_csv('taxonomy-table/taxonomy.tsv', sep='\t', header=0)
tax.head()

df = pd.read_csv('Songbird/amplitute-table/metadata.tsv', sep='\t', header=0)
df = df.drop(df.index[0])
df = pd.merge(df, tax, right_on='Feature ID', left_on = 'featureid', how = 'left')
df = df.drop(columns = 'featureid')
cols = df.columns.drop(['Feature ID','Taxon'])
df[cols] = df[cols].apply(pd.to_numeric, errors='coerce')
df[cols] = df[cols].round(3) 
df = df.iloc[(-df['AMAMPT'].abs()).argsort()]
df = df[:10]
df = df[['Feature ID', 'Taxon', 'AMAMPT']]
df

df.to_csv('amplitute.csv')

df = pd.read_csv('Songbird/amplitute-cat-table/metadata.tsv', sep='\t', header=0)
df = df.drop(df.index[0])
df = pd.merge(df, tax, right_on='Feature ID', left_on = 'featureid', how = 'left')
df = df.drop(columns = 'featureid')
cols = df.columns.drop(['Feature ID','Taxon'])
df[cols] = df[cols].apply(pd.to_numeric, errors='coerce')
df[cols] = df[cols].round(3) 
df = df.iloc[(-df['AMAMPT_C1'].abs()).argsort()]
df = df[:10]
df = df[['Feature ID', 'Taxon', 'AMAMPT_C1']]
df.to_csv('amplitute-cat.csv')
df.head()

df = pd.read_csv('Songbird/Acrophase-table/metadata.tsv', sep='\t', header=0)
df = df.drop(df.index[0])
df = pd.merge(df, tax, right_on='Feature ID', left_on = 'featureid', how = 'left')
df = df.drop(columns = 'featureid')
cols = df.columns.drop(['Feature ID','Taxon'])
df[cols] = df[cols].apply(pd.to_numeric, errors='coerce')
df[cols] = df[cols].round(3) 
df = df.iloc[(-df['AMPHIT'].abs()).argsort()]
df = df[:10]
df = df[['Feature ID', 'Taxon', 'AMPHIT']]
df.to_csv('Acrophase.csv')
df.head()

df = pd.read_csv('Songbird/AcrophaseSD-table/metadata.tsv', sep='\t', header=0)
df = df.drop(df.index[0])
df = pd.merge(df, tax, right_on='Feature ID', left_on = 'featureid', how = 'left')
df = df.drop(columns = 'featureid')
cols = df.columns.drop(['Feature ID','Taxon'])
df[cols] = df[cols].apply(pd.to_numeric, errors='coerce')
df[cols] = df[cols].round(3) 
df = df.iloc[(-df['AMPHIT_15SD'].abs()).argsort()]
df = df[:10]
df = df[['Feature ID', 'Taxon', 'AMPHIT_15SD']]
df.to_csv('AcrophaseSD.csv')
df.head()

df = pd.read_csv('Songbird/Fstats-cat-table/metadata.tsv', sep='\t', header=0)
df = df.drop(df.index[0])
df = pd.merge(df, tax, right_on='Feature ID', left_on = 'featureid', how = 'left')
df = df.drop(columns = 'featureid')
cols = df.columns.drop(['Feature ID','Taxon'])
df[cols] = df[cols].apply(pd.to_numeric, errors='coerce')
df[cols] = df[cols].round(3) 
df = df.iloc[(-df['AMFVT_C1'].abs()).argsort()]
df = df[:10]
df = df[['Feature ID', 'Taxon', 'AMFVT_C1']]
df.to_csv('Fstats-cat.csv')
df.head()

df = pd.read_csv('Songbird/Fstats-table/metadata.tsv', sep='\t', header=0)
df = df.drop(df.index[0])
df = pd.merge(df, tax, right_on='Feature ID', left_on = 'featureid', how = 'left')
df = df.drop(columns = 'featureid')
cols = df.columns.drop(['Feature ID','Taxon'])
df[cols] = df[cols].apply(pd.to_numeric, errors='coerce')
df[cols] = df[cols].round(3) 
df = df.iloc[(-df['AMFVT'].abs()).argsort()]
df = df[:10]
df = df[['Feature ID', 'Taxon', 'AMFVT']]
df.to_csv('Fstats.csv')
df.head()

df = pd.read_csv('Songbird/Poorslp-table/metadata.tsv', sep='\t', header=0)
df = df.drop(df.index[0])
df = pd.merge(df, tax, right_on='Feature ID', left_on = 'featureid', how = 'left')
df = df.drop(columns = 'featureid')
cols = df.columns.drop(['Feature ID','Taxon'])
df[cols] = df[cols].apply(pd.to_numeric, errors='coerce')
df[cols] = df[cols].round(3) 
df = df.iloc[(-df['PQBADSLP[T.1: Yes]'].abs()).argsort()]
df = df[:10]
df = df[['Feature ID', 'Taxon', 'PQBADSLP[T.1: Yes]']]
df.to_csv('Poorslp.csv')
df.head()

df = pd.read_csv('Songbird/Slphours-table/metadata.tsv', sep='\t', header=0)
df = df.drop(df.index[0])
df = pd.merge(df, tax, right_on='Feature ID', left_on = 'featureid', how = 'left')
df = df.drop(columns = 'featureid')
cols = df.columns.drop(['Feature ID','Taxon'])
df[cols] = df[cols].apply(pd.to_numeric, errors='coerce')
df[cols] = df[cols].round(3) 
df = df.iloc[(-df['SLEEPHRS[T.3:>=8 HRS]'].abs()).argsort()]
df = df[:10]
df = df[['Feature ID', 'Taxon', 'SLEEPHRS[T.3:>=8 HRS]']]
df.to_csv('Slphours.csv')
df.head()

df = pd.read_csv('Songbird/Slphours-con-table/metadata.tsv', sep='\t', header=0)
df = df.drop(df.index[0])
df = pd.merge(df, tax, right_on='Feature ID', left_on = 'featureid', how = 'left')
df = df.drop(columns = 'featureid')
cols = df.columns.drop(['Feature ID','Taxon'])
df[cols] = df[cols].apply(pd.to_numeric, errors='coerce')
df[cols] = df[cols].round(3) 
df = df.iloc[(-df['SLSLPHRS'].abs()).argsort()]
df = df[:10]
df = df[['Feature ID', 'Taxon', 'SLSLPHRS']]
df.to_csv('Slphours-con.csv')
df.head()
