==> 4.6 Differential Analysis -- Taxonomy significance (R).ipynb <==
setwd("~/Dropbox/lab/sleep/Results")
source('4.5 Differential Analysis -- Taxa test functions.R')
options(warn=-1)

df <- read.delim("~/Dropbox/lab/sleep/Results/mapping_sleep_alpha_omit_sleep.txt")
var.test <- c("AMFVT_C1", "AMAMPT_C1", "PQBADSLP", "SLEEPHRS",
              "AMFVT", "PQPEFFCY", "AMAMPT", "AMPHIT", "PQPSQI", "SLSLPHRS")
df <- df[ ,c("X.SampleID", var.test)]

file.name <- c("sample_plot_data.tsv", "sample_plot_data (1).tsv", "sample_plot_data (2).tsv",
               "sample_plot_data (3).tsv", "sample_plot_data (4).tsv", "sample_plot_data (5).tsv",
               "sample_plot_data (6).tsv", "sample_plot_data (7).tsv", "sample_plot_data (8).tsv",
               "sample_plot_data (9).tsv")

Acrophase <- get.pvalue(df, "Acrophase", file.name, var.test)
sum(Acrophase$AMPHIT < 0.05)

amplitude <- get.pvalue(df, "amplitude", file.name, var.test)
sum(amplitude$AMAMPT < 0.05)

amplitude_cat <- get.pvalue(df, "amplitude-cat", file.name, var.test)
sum(amplitude_cat$AMAMPT_C1 < 0.05)

Fstats <- get.pvalue(df, "Fstats", file.name,  var.test)
sum(Fstats$AMFVT < 0.05)

Fstats_cat <- get.pvalue(df, "Fstats-cat", file.name, var.test)
sum(Fstats_cat$AMFVT_C1 < 0.05)

Poorslp <- get.pvalue(df, "Poorslp", file.name,  var.test)
sum(Poorslp$PQBADSLP < 0.05)

Fstats

Fstats.adjust <- sapply(1:10, function(i){
  p.adjust(Fstats[,i], "BH")
})
colnames(Fstats.adjust) <- colnames(Fstats)[1:10]
as.data.frame(Fstats.adjust, colnames(Fstats.adjust))

Fstats_cat

Fstats_cat.adjust <- sapply(1:10, function(i){
  p.adjust(Fstats_cat[,i], "BH")
})
colnames(Fstats_cat.adjust) <- colnames(Fstats_cat)[1:10]
as.data.frame(Fstats_cat.adjust, colnames(Fstats_cat.adjust))
